Adding jest user and notification tests

+++ src/components/custom/templates/notification-page-template.tsx
@@ -30,7 +30,7 @@ export const NotificationPageTemplate = ({
 
   const handleDeleteAllNotifications = async () => {
     setDeleteLoading(true);
-    const res = await deleteAllNotificationsForUser(userId);
+    const res = await deleteAllNotificationsForUser({ userId: userId });
     if (res.success) {
       router.refresh();
     } else {
